Refactor web chat session controller
