+++ apps/web/src/chat/ChatPanel/tests/ChatPanel.post-run-sync.test.tsx
@@ -16,7 +16,7 @@ import {
 import {
   loadChatSessionWarmStartSnapshot,
   storeChatSessionWarmStartSnapshot,
-} from "../../sessionController/warmStart";
+} from "../../sessionController/lifecycle/warmStart";
 
 const {
   flushAsync,
