--- apps/web/src/chat/sessionController/snapshot.ts
@@ -1,3 +1,0 @@
-import type { ChatSessionSnapshot as ApiChatSessionSnapshot } from "../../types";
-
-export type ChatSessionSnapshot = ApiChatSessionSnapshot;
