+++ apps/web/src/chat/sessionController/state/snapshot.ts
@@ -1,0 +1,3 @@
+import type { ChatSessionSnapshot as ApiChatSessionSnapshot } from "../../../types";
+
+export type ChatSessionSnapshot = ApiChatSessionSnapshot;
